+++ truffle/CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 
 * GR-39219 Removed the deprecated `FrameSlot` API. The API was deprecated in 22.0.
 * GR-39219 Removed deprecated `CompilerOptions` API. The API was deprecated in 22.1.
+* GR-39219 Removed deprecated `TruffleRuntime.createCallTarget` and `RootNode.setCallTarget` API. The API was deprecated in 22.0.
 
 ## Version 22.2.0